fix(review): split quote/line coupling helpers out of validate-findings.mjs
validate-findings.mjs crossed its module-size budget (1057 vs. 985) after
merging upstream/main. Extract the quote/line coupling check --
quotableLines, quoteAppearsIn, lineIsAdded and addedLinesMatching -- into a
new sibling module, scripts/review/quote-line-coupling.mjs, and re-export
them from validate-findings.mjs so every existing import path (this file's
own checkProofOfWork/validateFindings, lane-canary.test.mjs, and
validate-findings.test.mjs) is unchanged.

validate-findings.mjs is now 967 lines (985 budget); the new module is 116
lines and needs no allowlist entry.
